unified-xai-platform/requirements.txt
# Unified XAI Platform - Requirements (Python 3.11 compatible)

# GUI Framework
streamlit>=1.28.0

# Deep Learning
tensorflow>=2.15.0
tf-keras>=2.15.0  # Keras 2 API for legacy model compatibility

# Audio Processing
librosa>=0.10.0
soundfile>=0.12.1

# Image Processing
opencv-python>=4.8.0
Pillow>=10.0.0

# XAI Libraries
lime>=0.2.0.1
shap>=0.43.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scikit-image>=0.21.0
